feat(OY2-26082): package actions multi-file download
