+++ src/services/ui/src/pages/detail/admin-changes/AdminChanges.tsx
@@ -60,7 +60,7 @@ export const AdminChange: FC<opensearch.changelog.Document> = (props) => {
         <p className="flex flex-row gap-2 text-gray-600">
           <strong>{label as string}</strong>
           {" - "}
-          {format(new Date(props.timestamp), "eee, MMM d, yyyy hh:mm:ss a OOO")}
+          {format(new Date(props.timestamp), "eee, MMM d, yyyy hh:mm:ss a")}
         </p>
       </AccordionTrigger>
       <AccordionContent className="p-4">
